COMP: Consistently rename "Method" and "utils" packages to avoid name conflicts
