+++ CMakeLists.txt
@@ -35,8 +35,7 @@ add_subdirectory(AREG)
 add_subdirectory(AutoMatrix)
 add_subdirectory(AutoCrop3D)
 add_subdirectory(FlexReg)
-add_subdirectory(Flex_Reg_CLI)
-
+add_subdirectory(FlexReg_CLI)
 add_subdirectory(MRI2CBCT)
 add_subdirectory(MRI2CBCT_CLI)
 ## NEXT_MODULE
