+++ FlexReg_CLI/FlexReg_CLI.xml
@@ -2,7 +2,7 @@
 <executable>
   <category>Automated Dental Tools.Advanced</category>
   <index>0</index>
-  <title>Flex_Reg_CLI</title>
+  <title>FlexReg_CLI</title>
   <description><![CDATA[CLI module that can be built into an extension]]></description>
   <version>0.1.0.</version>
   <documentation-url>https://github.com/GaelleLeroux/DCBIA_FlexReg</documentation-url>
