+++ FlexReg_CLI/__init__.py
@@ -1,0 +1,1 @@
+import FlexReg_CLI
