--- Flex_Reg_CLI/__init__.py
@@ -1,1 +1,0 @@
-import Flex_Reg_CLI
